rename the detector
