+++ mmdet/models/detectors/__init__.py
@@ -5,7 +5,7 @@ from .base import BaseDetector
 from .base_detr import DetectionTransformer
 from .cascade_rcnn import CascadeRCNN
 from .centernet import CenterNet
-from .conditioanl_detr import ConditionalDETR
+from .conditional_detr import ConditionalDETR
 from .cornernet import CornerNet
 from .ddod import DDOD
 from .deformable_detr import DeformableDETR
