+++ mmdet/models/detectors/conditional_detr.py
@@ -1,14 +1,15 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from typing import Dict, Tuple
+from torch import Tensor
+from .detr import DETR
 
 import torch
 import torch.nn as nn
-from torch import Tensor
+
+from ..layers import (DetrTransformerEncoder, ConditionalDetrTransformerDecoder,
+                      SinePositionalEncoding)
 
 from mmdet.registry import MODELS
-from ..layers import (ConditionalDetrTransformerDecoder,
-                      DetrTransformerEncoder, SinePositionalEncoding)
-from .detr import DETR
 
 
 @MODELS.register_module()
@@ -20,6 +21,9 @@ class ConditionalDETR(DETR):
     Code is modified from the `official github repo
     <https://github.com/Atten4Vis/ConditionalDETR>`_.
     """
+    def __init__(self, *arg, group_detr=1, **kwargs) -> None:
+        self.group_detr = group_detr
+        super().__init__(*arg, **kwargs)
 
     def _init_layers(self) -> None:
         """Initialize layers except for backbone, neck and bbox_head."""
@@ -31,11 +35,11 @@ class ConditionalDETR(DETR):
         # NOTE The embed_dims is typically passed from the inside out.
         # For example in DETR, The embed_dims is passed as
         # self_attn -> the first encoder layer -> encoder -> detector.
-        self.query_embedding = nn.Embedding(self.num_queries, self.embed_dims)
+        self.query_embedding = nn.Embedding(self.num_query * self.group_detr, self.embed_dims)
 
         num_feats = self.positional_encoding.num_feats
         assert num_feats * 2 == self.embed_dims, \
-            f'embed_dims should be exactly 2 times of num_feats. ' \
+            'embed_dims should be exactly 2 times of num_feats. ' \
             f'Found {self.embed_dims} and {num_feats}.'
 
     def pre_decoder(self, memory: Tensor) -> Tuple[Dict, Dict]:
@@ -68,8 +72,8 @@ class ConditionalDETR(DETR):
         if self.training:
             query_pos = self.query_embedding.weight
         else:
-            query_pos = self.query_embedding.weight[:self.num_queries]
-        # (num_queries, dim) -> (num_queries, bs, dim)
+            query_pos = self.query_embedding.weight[:self.num_query]
+        # (num_query, dim) -> (num_query, bs, dim)
         query_pos = query_pos.unsqueeze(1).repeat(1, batch_size, 1)
         query = torch.zeros_like(query_pos)
 
@@ -78,15 +82,19 @@ class ConditionalDETR(DETR):
         head_inputs_dict = dict()
         return decoder_inputs_dict, head_inputs_dict
 
-    def forward_decoder(self, query: Tensor, query_pos: Tensor, memory: Tensor,
-                        memory_mask: Tensor, memory_pos: Tensor) -> Dict:
+    def forward_decoder(self,
+                        query: Tensor,
+                        query_pos: Tensor,
+                        memory: Tensor,
+                        memory_mask: Tensor,
+                        memory_pos: Tensor) -> Dict:
         """Forward with Transformer decoder.
 
         Args:
             query (Tensor): The queries of decoder inputs, has shape
-                (num_queries, bs, dim).
+                (num_query, bs, dim).
             query_pos (Tensor): The positional queries of decoder inputs,
-                has shape (num_queries, bs, dim).
+                has shape (num_query, bs, dim).
             memory (Tensor): The output embeddings of the Transformer encoder,
                 has shape (num_feat, bs, dim).
             memory_mask (Tensor): ByteTensor, the padding mask of the memory,
@@ -96,10 +104,10 @@ class ConditionalDETR(DETR):
 
         Returns:
             dict: The dictionary of decoder outputs, which includes the
-            `hidden_states` and `references` of the decoder output.
+            `hidden_states` of the decoder output.#TODO
         """
-        # (num_decoder_layers, num_queries, bs, dim)
-        hidden_states, references = self.decoder(
+        # (num_decoder_layers, num_query, bs, dim)
+        hidden_states, reference_points = self.decoder(
             query=query,
             key=memory,
             value=memory,
@@ -107,6 +115,7 @@ class ConditionalDETR(DETR):
             key_pos=memory_pos,
             key_padding_mask=memory_mask)
         hidden_states = hidden_states.transpose(1, 2)
-        head_inputs_dict = dict(
-            hidden_states=hidden_states, references=references)
+        head_inputs_dict = dict(hidden_states=hidden_states, reference_points=reference_points)
         return head_inputs_dict
+
+
